pesel as bigint on sql

+++ CoursesAPI/Models/User.cs
@@ -9,7 +9,7 @@ namespace CoursesAPI.Models
         public string GivenName { get; set; }
         [Column(TypeName = "nvarchar(50)")]
         public string FamilyName { get; set; }
-        [Column(TypeName = "integer")]
+        [Column(TypeName = "bigint")]
         public long PeselNumber { get; set; }
         public string? RefreshToken { get; set; }
         public DateTime RefreshTokenExpiryTime { get; set; }
